requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.41.2
praw==7.7.1
PyYAML==6.0.1
Jinja2==3.1.3
python-dateutil==2.9.0
aiohttp==3.9.3
tenacity==8.2.3
rich==13.7.0
click==8.1.7
